requirements.txt
pytest==4.4.1
black==19.3b0
flake8==3.7.7
Django==2.1.4
Pillow==5.3.0
pytest-django==3.4.5
djangorestframework==3.9.3